Move tests around and clean up unused deps
