+++ graql/executor/BUILD
@@ -23,11 +23,12 @@ load("@graknlabs_build_tools//checkstyle:rules.bzl", "checkstyle_test")
 java_library(
     name = "executor",
     srcs = glob([
-        "**/*.java"
+        "**/*.java",
     ]),
+    tags = ["maven_coordinates=io.grakn.core:grakn-graql-executor:{pom_version}"],
     deps = [
         # Internal dependencies
-        "//common:common",
+        "//common",
         "//kb/concept/api",
         "//kb/concept/manager",
         "//kb/server",
@@ -37,16 +38,13 @@ java_library(
         "//kb/graql/planning",
         "//kb/keyspace",
         "//concept/answer",
-        "//concept/impl:impl",
-        "//concept/util",
         "//graql/analytics",
-        "//graql/reasoner", # TODO should this be pointing the other way?
+        "//graql/reasoner",  # TODO should this be pointing the other way?
         "//graql/planning",
         "//core",
 
         # External dependencies from @graknlabs
         "@graknlabs_graql//java:graql",
-        "@graknlabs_common//:common",
         "@graknlabs_benchmark//lib:server-instrumentation",
 
         # External dependencies from Maven
@@ -72,8 +70,6 @@ java_library(
         "//dependencies/maven/artifacts/org/scala-lang:scala-library",
         "//dependencies/maven/artifacts/org/slf4j:slf4j-api",
     ],
-    tags = ["maven_coordinates=io.grakn.core:grakn-graql-executor:{pom_version}"],
-
 )
 
 checkstyle_test(
