+++ test-integration/graql/executor/BUILD
@@ -25,14 +25,13 @@ java_test(
     classpath_resources = ["//test-integration/resources:logback-test"],
     test_class = "grakn.core.graql.executor.IsaExplicitIT",
     deps = [
+        "//common",
         "//dependencies/maven/artifacts/com/google/guava",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
-        "//graql/executor",
         "//graql/planning",
-        "//kb/server",
-        "//kb/graql/planning",
         "//kb/concept/api",
-        "//common",
+        "//kb/graql/planning",
+        "//kb/server",
         "//test-integration/rule:grakn-test-server",
         "@graknlabs_graql//java:graql",
     ],
