+++ test-integration/graql/planning/BUILD
@@ -28,13 +28,11 @@ java_test(
     deps = [
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
         "//dependencies/maven/artifacts/org/mockito:mockito-core",
-        "//graql/planning",
         "//graql/executor",
-        "//kb/server",
+        "//graql/planning",
         "//kb/concept/api",
         "//kb/concept/manager",
         "//kb/graql/planning",
-        "//server",
         "@graknlabs_graql//java:graql",
     ],
 )
@@ -48,17 +46,17 @@ java_test(
     ],
     test_class = "grakn.core.graql.gremlin.GraqlTraversalIT",
     deps = [
+        "//common",
         "//core",
         "//dependencies/maven/artifacts/com/google/guava",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
         "//dependencies/maven/artifacts/org/mockito:mockito-core",
-        "//graql/planning",
         "//graql/executor",
-        "//kb/server",
-        "//kb/graql/planning",
+        "//graql/planning",
         "//kb/concept/api",
         "//kb/concept/manager",
-        "//common",
+        "//kb/graql/planning",
+        "//kb/server",
         "//test-integration/rule:grakn-test-server",
         "@graknlabs_graql//java:graql",
     ],
