+++ test-integration/graql/query/BUILD
@@ -192,11 +192,10 @@ java_test(
         "//common",
         "//concept/impl",
         "//dependencies/maven/artifacts/com/google/guava",
+        "//graql/planning",
         "//kb/concept/api",
-        "//kb/graql/exception",
         "//kb/graql/planning",
         "//kb/server",
-        "//graql/planning",
         "//test-integration/rule:grakn-test-server",
         "//test-integration/util:concept-downcasting",
         "@graknlabs_graql//java:graql",
