+++ test-integration/graql/reasoner/query/BUILD
@@ -33,8 +33,8 @@ java_test(
         "//dependencies/maven/artifacts/com/google/guava",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
         "//graql/reasoner",
-        "//kb/graql/reasoner",
         "//kb/graql/exception",
+        "//kb/graql/reasoner",
         "//kb/server",
         "//test-integration/rule:grakn-test-server",
         "//test-integration/util:graql-test-util",
@@ -108,7 +108,6 @@ java_test(
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
         "//graql/reasoner",
         "//kb/concept/api",
-        "//kb/graql/reasoner",
         "//kb/graql/exception",
         "//kb/server",
         "//test-integration/graql/reasoner/graph:reachability-graph",
@@ -200,12 +199,12 @@ java_test(
         "//common",
         "//dependencies/maven/artifacts/com/google/guava",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
-        "//graql/reasoner",
         "//graql/planning",
+        "//graql/reasoner",
         "//kb/concept/api",
         "//kb/concept/manager",
-        "//kb/graql/reasoner",
         "//kb/graql/planning",
+        "//kb/graql/reasoner",
         "//kb/keyspace",
         "//kb/server",
         "//keyspace",
