+++ test-integration/keyspace/KeyspaceSchemaCacheIT.java
@@ -14,9 +14,10 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
  */
 
-package grakn.core.server.session.cache;
+package grakn.core.keyspace;
 
 import grakn.client.GraknClient;
 import grakn.core.kb.concept.api.Role;
