+++ test-integration/server/BUILD
@@ -65,10 +65,68 @@ java_test(
     ],
 )
 
+
+
+java_test(
+    name = "validator-it",
+    size = "medium",
+    srcs = ["ValidatorIT.java"],
+    classpath_resources = ["//test-integration/resources:logback-test"],
+    test_class = "grakn.core.server.kb.ValidatorIT",
+    deps = [
+        "//common:common",
+        "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
+        "//kb/server",
+        "//kb/concept/api",
+        "//test-integration/rule:grakn-test-server",
+        "@graknlabs_graql//java:graql",
+    ],
+)
+
+java_test(
+    name = "validate-global-rules-it",
+    size = "medium",
+    srcs = ["ValidateGlobalRulesIT.java"],
+    classpath_resources = ["//test-integration/resources:logback-test"],
+    test_class = "grakn.core.server.kb.ValidateGlobalRulesIT",
+    deps = [
+        "//concept/impl:impl",
+        "//kb/server",
+        "//kb/concept/api",
+        "//kb/concept/structure",
+        "//server",
+        "//test-integration/rule:grakn-test-server",
+        "//test-integration/util:concept-downcasting",
+    ],
+)
+
+java_test(
+    name = "rule-validation-it",
+    size = "large",
+    srcs = ["RuleValidationIT.java"],
+    classpath_resources = ["//test-integration/resources:logback-test"],
+    test_class = "grakn.core.server.kb.RuleValidationIT",
+    deps = [
+        "//common:common",
+        "//core",
+        "//dependencies/maven/artifacts/com/google/guava",
+        "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
+        "//kb/server",
+        "//kb/concept/api",
+        "//graql/reasoner",
+        "//test-integration/rule:grakn-test-server",
+        "@graknlabs_graql//java:graql",
+    ],
+)
+
+
 checkstyle_test(
     name = "checkstyle",
     targets = [
         ":attribute-uniqueness-it",
         ":grakn-client-it",
+        ":validator-it",
+        ":validate-global-rules-it",
+        ":rule-validation-it",
     ],
 )
