+++ test-integration/server/ValidateGlobalRulesIT.java
@@ -14,9 +14,10 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
  */
 
-package grakn.core.server.kb;
+package grakn.core.server;
 
 import grakn.core.concept.impl.RelationImpl;
 import grakn.core.kb.concept.api.Entity;
