--- test-integration/server/statistics/BUILD
@@ -1,58 +1,0 @@
-#
-# GRAKN.AI - THE KNOWLEDGE GRAPH
-# Copyright (C) 2019 Grakn Labs Ltd
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-
-load("@graknlabs_build_tools//checkstyle:rules.bzl", "checkstyle_test")
-
-java_test(
-    name = "statistics-delta-it",
-    srcs = ["StatisticsDeltaIT.java"],
-    classpath_resources = ["//test-integration/resources:logback-test"],
-    test_class = "grakn.core.server.statistics.StatisticsDeltaIT",
-    deps = [
-        "//kb/concept/api",
-        "//kb/keyspace",
-        "//kb/server",
-        "//test-integration/rule:grakn-test-server",
-        "@graknlabs_graql//java:graql",
-    ],
-)
-
-java_test(
-    name = "keyspace-statistics-it",
-    srcs = ["KeyspaceStatisticsIT.java"],
-    classpath_resources = ["//test-integration/resources:logback-test"],
-    test_class = "grakn.core.server.statistics.KeyspaceStatisticsIT",
-    deps = [
-        "//concept/answer",
-        "//core",
-        "//kb/concept/api",
-        "//kb/keyspace",
-        "//kb/server",
-        "//test-integration/rule:grakn-test-server",
-        "@graknlabs_client_java//:client-java",
-        "@graknlabs_graql//java:graql",
-    ],
-)
-
-checkstyle_test(
-    name = "checkstyle",
-    targets = [
-        ":statistics-delta-it",
-        ":keyspace-statistics-it",
-    ],
-)
